ok, got bigendian byte order requests working.

+++ source/include/rpc_lsa.h
@@ -201,7 +201,7 @@ typedef struct r_lsa_query_sec_obj_info
 typedef struct lsa_query_info
 {
 	POLICY_HND pol; /* policy handle */
-	uint8 info_class; /* info class */
+	uint16 info_class; /* info class */
 
 } LSA_Q_QUERY_INFO;
